+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+### Fixed
+ - Fix Protontricks crash when custom Proton has an empty `compatibilitytool.vdf` manifest
+
 ## [1.11.0] - 2023-12-30
 ### Added
  - Show app icons for custom shortcuts in the app selector